+++ osrframework/utils/general.py
@@ -127,7 +127,7 @@ def _generateTabularData(res, oldTabularData = {}, isTerminal=False, canUnicode=
         return h
         
     # Entities allowed for the output in terminal
-    allowedInTerminal = ["i3visio_alias", "i3visio_uri", "i3visio_platform", "i3visio_email", "i3visio_ipv4", "i3visio_phone"]
+    allowedInTerminal = ["i3visio_alias", "i3visio_uri", "i3visio_platform", "i3visio_email", "i3visio_ipv4", "i3visio_phone", "i3visio_dni"]
     # List of profiles found
     values = {}
     headers = ["_id"]